rename files to match format
