+++ shared/domain/src/commonMain/kotlin/com/android254/shared/domain/models/SessionDomainModel.kt
@@ -15,16 +15,23 @@
  */
 package com.android254.shared.domain.models
 
-data class HomeDetails(
-    val isCallForSpeakersEnable: Boolean,
-    val linkToCallForSpeakers: String,
-    val isEventBannerEnable: Boolean,
-    val sessions: List<Session>,
-    val sessionsCount: Int,
-    val isSessionsSectionEnable: Boolean,
-    val speakers: List<Speaker>,
-    val speakersCount: Int,
-    val isSpeakersSessionEnable: Boolean,
-    val sponsors: List<Sponsors>,
-    val organizers: List<OrganizingPartners>
+data class SessionDomainModel(
+    val id: String,
+    val endDateTime: String,
+    val endTime: String,
+    val isBookmarked: Boolean,
+    val isKeynote: Boolean,
+    val isServiceSession: Boolean,
+    val sessionImage: String?,
+    val startDateTime: String,
+    val startTime: String,
+    val rooms: String,
+    val speakers: List<SpeakerDomainModel>,
+    val remoteId: String,
+    val description: String,
+    val sessionFormat: String,
+    val sessionLevel: String,
+    val slug: String,
+    val title: String,
+    val eventDay: String
 )
